+++ src/lib/pages/SaWTab.svelte
@@ -1,0 +1,40 @@
+<script>
+    // Images
+    import optkkunUrl from '$lib/optkkun.png';
+</script>
+
+<!-- Links ImageSrc={[file-name]ScreenshotUrl} -->
+    
+<h1>Socials and Websites</h1>
+<div class="container">
+    <ul>
+        <h2>Socials</h2>
+        <li><a a href="https://discord.gg/jvmadKWFe9" target='_blank' class='text-blue-600 underline'>Official English Discord Server</a> <i class="fa-solid fa-arrow-up-right-from-square text-blue-600"></i><br></li>
+        <li><a a href="https://discord.gg/CJ4nTkpy7t" target='_blank' class='text-blue-600 underline'>Official Japanese Discord Server</a> <i class="fa-solid fa-arrow-up-right-from-square text-blue-600"></i><br></li>
+        <li><a a href="https://www.youtube.com/@OpenTaiko" target='_blank' class='text-blue-600 underline'>Official YouTube Channel</a> <i class="fa-solid fa-arrow-up-right-from-square text-blue-600"></i><br></li>
+        <li><a a href="https://bsky.app/profile/opentaiko.bsky.social" target='_blank' class='text-blue-600 underline'>Official Bluesky Account</a> <i class="fa-solid fa-arrow-up-right-from-square text-blue-600"></i><br></li>
+        <li><a a href="https://x.com/OpenTaiko" target='_blank' class='text-blue-600 underline'>Official Twitter/X Account</a> <i class="fa-solid fa-arrow-up-right-from-square text-blue-600"></i><br></li>
+        <li><a a href="https://github.com/OpenTaiko" target='_blank' class='text-blue-600 underline'>GitHub Organization</a> <i class="fa-solid fa-arrow-up-right-from-square text-blue-600"></i><br></li>
+    </ul>
+    <img src={optkkunUrl} alt="OpenTaiko-kun" class="image optkkun"/>
+    <ul>
+        <h2>Websites</h2>
+        <li><a a href="https://opentaiko.github.io/" target='_blank' class='text-blue-600 underline'>Official OpenTaiko Website</a> <i class="fa-solid fa-arrow-up-right-from-square text-blue-600"></i><br></li>
+        <li><a a href="https://opentaiko.wiki.gg/" target='_blank' class='text-blue-600 underline'>Official OpenTaiko Wiki</a> <i class="fa-solid fa-arrow-up-right-from-square text-blue-600"></i><br></li>
+    </ul>
+</div>
+
+<style>
+    .container{
+        display: flex;
+        align-items:center;
+    }
+    ul.a {
+        list-style-type: circle;
+    }
+    img.optkkun {
+        display: block;
+        margin-left: auto;
+        margin-right: 60px;
+    }
+</style>