+++ CHANGELOG.md
@@ -3,6 +3,15 @@
 You can update your OpenTaiko Hub version directly from the releases tab:
 - https://github.com/OpenTaiko/OpenTaiko-Hub/releases
 
+## [0.1.6] - 2025-07-?? (Alpha)
+
+- [Feat] Added the "Socials and Websites" tab.
+- [Feat] Added Subtitle Edit to the "Charting" tab under "Tools".
+- [Chore] Changed the GitHub link to go to the OpenTaiko Hub's repository.
+- [Chore] Moved the Discord link to the "Socials and Websites" tab.
+- [Chore] Updated the HoF crowns to link to https://opentaiko.github.io/.
+- [Chore] Fixed some typos.
+
 ## [0.1.5] - 2025-05-02 (Alpha)
 
 - [Fix] AppImage should now be fully functional on Linux